+++ frontend/src/common/ui-components/common/Page/Page.module.scss
