+++ frontend/src/common/ui-components/common/styles/LoadingBar.module.scss
