+++ frontend/src/common/ui-components/common/styles/LogWrapper.module.scss
