+++ frontend/src/modules/common/Error/ErrorResponseWrapper.tsx
@@ -1,4 +1,5 @@
 import React from "react";
+// eslint-disable-next-line css-modules/no-unused-class
 import styles from "../Error/ErrorStatusWrapper.module.scss";
 import { ResponseStatusError } from "../../Nodes/context/NodesContext";
